demo1/requirements.txt
# AutoGen 多智能体框架
autogen-agentchat==0.4.2
autogen-ext[openai]==0.4.2

# 环境变量管理
python-dotenv==1.0.0

# OpenAI SDK
openai>=1.12.0

# 智谱AI SDK（用于联网搜索）
zai-sdk>=0.2.0

# 终端UI
rich>=13.0.0

# 其他依赖
pydantic>=2.0.0
